// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  brand: typeof brand;
  http: typeof http;
  integrations: typeof integrations;
  outputs: typeof outputs;
  posts: typeof posts;
  repositories: typeof repositories;
  webhook_logs: typeof webhook_logs;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;